Added -U on pip install
On Ubuntu's system, the `--user` is irrelevant (since that's the default behavior), but on other Linux systems using pip on a non virtual environment and without that flag will try to install the packages system-wide and will fail because of the lack of sudo permissions. Adding the `--user` makes this a more general script that might be used for other Linux systems.